requirements.txt
aioboto3==13.1.1
aiobotocore==2.13.1
aiofiles==24.1.0
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
anyio==4.4.0
astrapy==1.4.0
attrs==23.2.0
boto3==1.34.131
botocore==1.34.131
bson==0.5.10
cassandra-driver==3.29.1
cassio==0.1.8
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.5.14
deeplake==3.9.12
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
frozenlist==1.4.1
geomet==0.2.1.post1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
humbug==0.3.2
hyperframe==6.0.1
idna==3.7
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.7
langchain-astradb==0.3.3
langchain-community==0.2.7
langchain-core==0.2.16
langchain-openai==0.1.15
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langchainplus-sdk==0.0.20
langsmith==0.1.85
libdeeplake==0.0.134
lz4==4.3.3
marshmallow==3.21.3
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numexpr==2.10.1
numpy==1.26.4
openai==1.35.13
openapi-schema-pydantic==1.2.4
orjson==3.10.6
packaging==24.1
pathos==0.3.2
pillow==10.2.0
pox==0.3.4
ppft==1.7.6.8
pydantic==1.10.17
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
tenacity==8.5.0
tiktoken==0.7.0
toml==0.10.2
tqdm==4.66.4
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
uuid6==2024.1.12
wrapt==1.16.0
yarl==1.9.4
